dom0_construct_pv: set SIF_INITDOMAIN iff is_control_domain
This will affect the behaviour of the PV hardware domain that
we're currently constructing via dom0 logic, since it changes the
flags exposed to the guest, and Linux has logic gated on the
"xen_initial_domain" predicate that this flag determines.

Unclear whether we want or need this so far; may be reverted.