introduce hyperlaunch kconfig and core
Add a CONFIG_HYPERLAUNCH to allow enabling/disabling the hyperlaunch code
paths. Introduce an initial use of the hyperlaunch_enabled variable to suppress
the warning about multiple initrd candidates when multiboot modules are
detected when the hyperlaunch boolean is true.

Added to common code since this option will apply to all architectures.

Signed-off-by: Christopher Clark <christopher.clark@starlab.io>
Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>