xsm: refactor flask sid alloc and domain check
The function flask_domain_alloc_security() is where a default SID should be
assigned to a domain under construction. For reasons unknown, the initial
domain would be assigned unlabeled_t and then fixed up under
flask_domain_create().  With the introduction of xenboot_t it is now possible
to distinguish when the hypervisor is in the boot state.

This commit addresses the above situation by using a check to confirm if the
hypervisor is under the xenboot_t context in flask_domain_alloc_security(). If
that is the case, then it will inspect the domain's is_privileged field, and
select the appropriate default label, dom0_t or domU_t, for the domain. The
logic for flask_domain_create() was changed to allow the incoming SID to
override the default label.

The base policy was adjusted to allow the idle domain under the xenboot_t
context to be able to construct domains of both types, dom0 and domU.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>